pyproject.toml
[tool.poetry]
name = "hiking"
version = "0.1.0"
description = ""
authors = ["hunter britten <huntergbritten@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
djangorestframework = "^3.12.4"
docker-compose = "^1.29.2"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
